+++ temporal-sdk/src/main/java/io/temporal/internal/common/ProtoEnumNameUtils.java
@@ -23,6 +23,7 @@ package io.temporal.internal.common;
 import com.google.common.base.CaseFormat;
 import com.google.common.collect.ImmutableMap;
 import io.temporal.api.enums.v1.CommandType;
+import io.temporal.api.enums.v1.EventType;
 import io.temporal.api.enums.v1.IndexedValueType;
 import io.temporal.api.enums.v1.WorkflowTaskFailedCause;
 import java.util.Map;
@@ -51,12 +52,18 @@ public class ProtoEnumNameUtils {
   public static final String WORKFLOW_TASK_FAILED_CAUSE_PREFIX = "WORKFLOW_TASK_FAILED_CAUSE_";
   public static final String INDEXED_VALUE_TYPE_PREFIX = "INDEXED_VALUE_TYPE_";
   public static final String COMMAND_TYPE_PREFIX = "COMMAND_TYPE_";
+  public static final String EVENT_TYPE_PREFIX = "EVENT_TYPE_";
 
   public static final Map<Class<?>, String> ENUM_CLASS_TO_PREFIX =
       ImmutableMap.of(
-          WorkflowTaskFailedCause.class, WORKFLOW_TASK_FAILED_CAUSE_PREFIX,
-          IndexedValueType.class, INDEXED_VALUE_TYPE_PREFIX,
-          CommandType.class, COMMAND_TYPE_PREFIX);
+          WorkflowTaskFailedCause.class,
+          WORKFLOW_TASK_FAILED_CAUSE_PREFIX,
+          IndexedValueType.class,
+          INDEXED_VALUE_TYPE_PREFIX,
+          CommandType.class,
+          COMMAND_TYPE_PREFIX,
+          EventType.class,
+          EVENT_TYPE_PREFIX);
 
   @Nonnull
   public static String uniqueToSimplifiedName(@Nonnull Enum<?> enumm) {
